UI updates

+++ libqpamac/include/Utils.h
@@ -18,7 +18,7 @@ namespace Utils {
     Q_PROPERTY(QDateTime name READ name CONSTANT)\
     inline QDateTime name()\
 {\
-    return QDateTime::fromMSecsSinceEpoch(method);\
+    return QDateTime::fromMSecsSinceEpoch(qint64(method));\
     }
 #define PAMAC_QT_STRING_PROPERTY_GET_SET(getName,getMethod,setName,setMethod)\
     Q_PROPERTY(QString getName READ getName WRITE setName)\
